infra/init: manage infrastructure with terraform
apex has support out of the box for terraform so we should use it to
manage our AWS infrastructure. It gives us great granularity to manage
our services and we could build/find tools to make it super simple to
generate terrform configurations for the common tasks such as
provisioning endpoints.